# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vullify API
  description: |
    REST API for container registry integration, vulnerability scans (Trivy), findings,
    enrichment (EPSS / KEV), and reporting. List responses wrap payloads in `data` with optional `meta` pagination.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/scans:
    post:
      tags: [Scans]
      summary: Create scan
      description: Creates a scan job and enqueues it on Redis for the worker.
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScan"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/scans/{id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Get scan detail
      operationId: getScan
      responses:
        "200":
          description: Scan with severity counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/findings:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: List findings for scan
      operationId: listScanFindings
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: package_name
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingListEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/v1/scans/{id}/sbom:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Get raw SBOM JSON
      description: Returns JSON body; `X-SBOM-Format` header indicates format.
      operationId: getScanSBOM
      responses:
        "200":
          description: SBOM payload
          headers:
            X-SBOM-Format:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/findings/{id}:
    parameters:
      - $ref: "#/components/parameters/FindingId"
    get:
      tags: [Findings]
      summary: Get finding detail
      description: Includes enrichment fields when available (EPSS, KEV, known exploits).
      operationId: getFinding
      responses:
        "200":
          description: Finding detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingDetailEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/reports/vulnerability:
    get:
      tags: [Reports]
      summary: Vulnerability occurrence report
      operationId: vulnerabilityReport
      parameters:
        - name: severity
          in: query
          schema:
            type: string
        - name: from
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: End date (YYYY-MM-DD), inclusive end-of-day
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Paginated vulnerability rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VulnerabilityReportEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    ScanId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FindingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error:
              code: VALIDATION_ERROR
              message: invalid JSON body
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          $ref: "#/components/schemas/Meta"
...
    CreateScan:
      type: object
      description: Provide exactly one of image_id or image_ref
      properties:
        image_id:
          type: string
          format: uuid
        image_ref:
          type: string
          description: Canonical pull reference matching stored images
    ScanDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        status:
          type: string
        triggered_by:
          type: string
        started_at:
          type: string
          nullable: true
        completed_at:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        trivy_version:
          type: string
          nullable: true
        repository:
          type: string
        tag:
          type: string
        severity_counts:
          type: object
          additionalProperties:
            type: integer
            format: int64
    ScanDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ScanDetail"
    FindingListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vulnerability_id:
          type: string
        package_name:
          type: string
        severity:
          type: string
        created_at:
          type: string
        installed_version:
          type: string
        fixed_version:
          type: string
        title:
          type: string
    FindingListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/FindingListItem"
    FindingDetail:
      type: object
      description: Dynamic object; optional enrichment fields
      additionalProperties: true
    FindingDetailEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/FindingDetail"
    VulnerabilityReportRow:
      type: object
      properties:
        vulnerability_id:
          type: string
        severity:
          type: string
        occurrences:
          type: integer
        last_seen:
          type: string
    VulnerabilityReportEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/VulnerabilityReportRow"
...